// api/src/prompt/enums/prompt-message.enum.ts
export enum PromptMessageContentType {
    TEXT = 'text',
    IMAGE = 'image',
    AUDIO = 'audio',
    VIDEO = 'video',
    DOCUMENT = 'document',
}

export enum ImageDetail {
    LOW = 'low',
    HIGH = 'high',
}

export enum PromptMessageRole {
    SYSTEM = 'sytem',
    USER = 'user',
    ASSISTANT = 'assistant',
    TOOL = 'tool',
}